segmentation/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.1
selenium==4.17.2
webdriver-manager==4.0.2
Pillow==10.4.0
ImageHash==4.3.1
google-generativeai==0.4.1
